pyproject.toml
[tool.poetry]
name = "cv_project"
version = "0.1.0"
description = ""
authors = ["Divaltor <kuklavudu78@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.9"
Django = "^3.2.4"
djangorestframework = "^3.12.4"
envparse = "^0.2.0"
loguru = "^0.5.3"
httpx = "^0.18.1"
Pillow = "^8.2.0"
any-case = "^0.1.6"
celery = "5.0.5"
psycopg2-binary = "^2.8.6"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
